@@ -24,8 +24,8 @@
 #include <Writer.hpp>
 #include <Decoder.hpp>
 
-AVThread::AVThread(PlayClass &playC, const QString &writer_type, Writer *_writer, const QStringList &pluginsName) :
-    dec(nullptr), playC(playC), br(false), br2(false), waiting(false),
+AVThread::AVThread(PlayClass & __playC, const QString &writer_type, Writer *_writer, const QStringList &pluginsName) :
+    dec(nullptr), playC(__playC), br(false), br2(false), waiting(false),
     mutex(QMutex::Recursive)
 {
     connect(this, SIGNAL(finished()), this, SLOT(deleteLater()));
